pyproject.toml
[project]
name = "multi_rag"
version = "0.1.0"
description = "Multi Modal Rag"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "glmocr>=0.1.0",
    "pymupdf>=1.27.2",
    "Pillow>=12.1.1",
    "pyyaml>=6.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.8.0",
    "python-dotenv>=1.0.0",
    "rich>=14.0.0",
    "tqdm>=4.67.0",
    "streamlit>=1.40.0",
    "openai>=2.24.0",
    "qdrant-client>=1.17.0",
    "tiktoken>=0.9.0",
    "httpx>=0.28.0",
    "fastapi>=0.120.0",
    "uvicorn[standard]>=0.34.0",
    "loguru>=0.7.0",
    "python-multipart>=0.0.20",
    "opencv-python>=4.13.0.92",
    "pytest>=8.4.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.11.0",
    "mypy>=1.15.0",
    "FlagEmbedding>=1.3.0",
    "transformers>=4.51.0",
    "torch>=2.7.0",
    "google-genai>=1.0.0",
    "glmocr[layout]",
    "torch>=2.10",
    "torchvision>=0.25",
    "transformers>=5.3",
    "sentencepiece>=0.2",
    "accelerate>=1.13",
    "opencv-python>=4.10.0",
]
